The commit message for this git diff would be:
This commit message uses the conventional commit type "refactor" to indicate that changes were made to the codebase, but not to the content or behavior of the software itself. The message also includes a brief description of what was changed (updating lines added/removed/active time in README.md), which helps the developer understand the purpose of this commit.