Here's a possible commit message based on the provided git diff and modified files:
The commit message includes a short description of what changed (updating the statistics in the README file and adding a visualization) and why it was made (to provide a more accurate representation of our team's productivity and help us identify trends in our coding activity). The type of change is specified as "docs" to indicate that this is a documentation-related commit.