Here is a concise commit message based on the provided git diff:
The commit message starts with the conventional type , which indicates that this change is related to documentation. The subject line describes the actual changes made, in this case, updating the README file with new lines added and removed. The body of the commit message includes a list of modified files to provide more context about the changes made.