Here is a concise commit message based on the given git diff and modified files:
The commit message provides a clear and concise summary of the changes made, focusing on what has been modified and why. The  type is used because the changes do not affect the functionality of the application in any way, but rather updating some data. The message also includes a list of the affected files and the duration of the activity.