Here is a concise commit message based on the given git diff and modified files:
This message uses the conventional commit type  to indicate that the changes are not related to adding new features or fixing bugs, but rather a refactoring of existing code. The message also includes a brief description of the changes made in the report, specifically updating the coding activity for June 26th, 2025.
The  section lists the files that were modified, which include the  file and the  file within the  directory. The message also includes the  of 105 minutes to indicate how long it took to update the report.