The message for this commit would be:
This type of commit is used to indicate that a change has been made to the system or its configuration. In this case, the lines added, lines removed, and active time have been updated in the activity file, which is a necessary part of tracking coding activities.
The commit message should be brief and descriptive, focusing on what changed and why it was made. The conventional commit type, refactor, is used to indicate that this is an update rather than a new feature or bug fix.