The commit message for this git diff is:
This message includes a clear description of what changed (the update to the activity metrics) and why (to reflect the changes in the files). It also includes the conventional commit type "refactor" to indicate that this change does not add any new functionality, but rather simplifies the codebase by updating some existing data structures.