Here is a concise commit message based on the provided git diff and modified files:
This commit message uses the conventional commit type "refactor" to indicate that we are making changes to the codebase without adding new functionality or fixing bugs. The message summarizes the main changes made in the commit, focusing on WHAT changed and WHY. In this case, we optimized the performance of the  file by reducing the number of lines added and removed.