Here is a concise commit message based on the provided git diff and modified files:
This commit message follows the conventional commit format (refactor) and describes the main changes made to the activity report for 30th June 2025. The message highlights the updates to the 'activity.json' file, including the total lines added, removed, active time, and net change, as well as the addition of a new file ('cda/README.md') with updated information.