Here's a concise commit message based on the provided git diff and modified files:
In this message, we use the conventional commit type "fix" to indicate that we have made changes to our activity statistics. We also provide a brief summary of what has changed in the project, including the modified functions and files, as well as the duration of the activity. The message is clear and descriptive, focusing on what has changed and why.