Here's a concise commit message based on the provided git diff and modified files:
In this commit message, we use the conventional type "docs" to indicate that we are updating documentation. The scope is "cda", which refers to the project directory. The summary of the commit message provides a brief overview of the changes made in the README file. Specifically, we update the statistics and add helperText changes.
We can also summarize the main changes by listing the modified files: