This is a concise commit message based on the given git diff and modified files:
The commit type  indicates that this change is related to maintenance or setup tasks, rather than adding new features or fixing bugs. The scope of the commit is specific to the affected project ().
In the message body, we can find information about what changed:
* Updated project statistics and added a new file ()
We also mentioned why this change was necessary in the message body.