Here's a concise commit message based on the provided git diff and modified files:
This message describes the changes made in the commit, including the type of change (refactor), the affected files, and the duration of the activity. It also provides a brief summary of the main changes, which is "update total lines added and removed" in this case. The message includes the name of the modified functions and the files that were changed, as well as the activity duration.