Here's a concise commit message based on the given information:
In this message, we use the conventional commit type "refactor" to indicate that we have made changes to our codebase without adding new features or fixing bugs. We also provide a brief summary of what has changed and why we made these modifications. Specifically, we update the activity data in  and increase the total active time by 17 minutes.