The commit message for this set of changes would be:
This type of commit is a "chore" which indicates that the change does not affect the overall functionality or behavior of the code, but rather improves its maintenance and administration. The message describes the specific changes made to the activity statistics, including the addition of 156 lines removed, a net change of 17244, and an increase in active time by 5 minutes.