This commit message is:
This message follows the conventional commit format, which includes the following information:
* The type of change (in this case, a "feat" for a feature)
* A short description of the changes (in this case, updating activity statistics for a specific day)
* The modified files affected by the changes
* A list of changes made to the files
The message is clear and descriptive, providing an overview of the changes made on 2025-07-08 and highlighting the main changes and affected functionality. It also follows the conventional commit format, making it easy for others to understand the purpose and impact of the changes.